RavenDB-15348 "Your current license doesn't include the encryption feature" - more information for further tests failing investigation
